requirements.txt
boto3
google-api-python-client
google-auth
python-docx